Precompile heat folder
